Move HttpStatus to its own file
Apparently stuff breaks if common/http is used in the server.
